Fix clang-tidy findings in the compact symbols code
Give `Symbol::offset_begin`/`offset_end` default member initializers
(the tagged name reference made the implicit constructor subject to
cppcoreguidelines-pro-type-member-init), make `section_name` a plain
char array so it is usable as a C string without a `data()` call, store
the `ZSTD_getFrameContentSize` result in `uint64_t` (google-runtime-int)
and annotate the six `demangle`/`tryDemangle` calls on views returned by
`getSymbolName` (NUL-terminated by contract) with targeted NOLINT.

CI report: https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=115323&sha=e133ad047fc00b801d1a9a1ce457e623a0855a2a&name_0=PR&name_1=Build%20%28arm_tidy%29
Related: https://github.com/ClickHouse/ClickHouse/pull/115323

+++ src/Common/CompactSymbols.cpp
@@ -86,7 +86,7 @@ std::vector<char> compress(std::span<const char> input)
 
 size_t frameContentSize(std::string_view input)
 {
-    unsigned long long decompressed_size = ZSTD_getFrameContentSize(input.data(), input.size());
+    uint64_t decompressed_size = ZSTD_getFrameContentSize(input.data(), input.size());
     if (decompressed_size == ZSTD_CONTENTSIZE_ERROR)
         throw std::runtime_error("Invalid zstd frame in compact symbols");
     if (decompressed_size == ZSTD_CONTENTSIZE_UNKNOWN)
